refactor: оптимизация header, разделение на более мелкие компоненты
