+++ src/components/layout/header/components/favorite-count.tsx
@@ -1,0 +1,10 @@
+import { useAppSelector } from '../../../../hooks';
+import { offersSelectors } from '../../../../store/slices/offers';
+
+export default function FavoriteCount(): JSX.Element {
+  const favoritesOffersCount = useAppSelector(offersSelectors.favoritesOffersCount);
+
+  return (
+    <span className="header__favorite-count">{favoritesOffersCount}</span>
+  );
+}
