+++ src/components/layout/header/components/login.tsx
@@ -1,0 +1,15 @@
+import { AuthorizationStatus } from '../../../../const';
+import { useAppSelector } from '../../../../hooks';
+import { authorizationSelectors } from '../../../../store/slices/authorization';
+
+export default function Login(): JSX.Element {
+  const authorizationStatus = useAppSelector(authorizationSelectors.authorizationStatus);
+
+  return (
+    <span className="header__login">
+      {authorizationStatus === AuthorizationStatus.Unknown ?
+        'загрузка...' :
+        'Sign in'}
+    </span>
+  );
+}
