+++ src/components/layout/header/components/user-name.tsx
@@ -1,0 +1,12 @@
+
+import { useAppSelector } from '../../../../hooks';
+import { authorizationSelectors } from '../../../../store/slices/authorization';
+
+
+export default function UserName(): JSX.Element {
+  const email = useAppSelector(authorizationSelectors.email);
+
+  return (
+    <span className="header__user-name user__name">{email}</span>
+  );
+}
